requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.3
cfgv==3.4.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distlib==0.3.9
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
identify==2.6.12
kombu==5.5.4
nodeenv==1.9.1
packaging==25.0
platformdirs==4.3.8
pre_commit==4.2.0
prompt_toolkit==3.0.51
psycopg==3.2.9
psycopg-binary==3.2.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
redis==6.2.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
vine==5.1.0
virtualenv==20.31.2
wcwidth==0.2.13
